chore: share sample graph between ui-thread and web-worker variants, add README
Move the 5-node demo graph into src/shared/example-graph.js so ui-thread
and web-worker variants differ only in where the router runs. Document
all three variants in js/README.md.
