+++ libavoid-standalone-link-routing/js/README.md
@@ -31,8 +31,32 @@ Install the dependencies by running:
 npm install
 ```
 
-And then start the application with:
+And then start one of the variants below.
+
+## Variants
+
+Three variants live side-by-side under `src/`, sharing the avoid-router integration code (and the sample graph used by the first two) in `src/shared/`. They differ in *where* the router runs and *what* each one is meant to demonstrate.
+
+### UI thread (default)
+
+Runs the avoid router on the main thread against a small graph (5 nodes, 4 links) built by `src/shared/example-graph.js`. Full interactive editing: double-click blank to add a node, drag to move or resize, hover a link for remove/arrowhead tools. The simplest place to read the router integration end to end.
 
 ```bash
-npm run start
+npm run start   # or: npm run build
+```
+
+### Web Worker
+
+Same graph, same shapes, same editing affordances as the UI-thread variant — the only meaningful difference is that the avoid router runs in a dedicated `Worker` (`src/web-worker/worker.js`) so routing never blocks interaction. Demonstrates the full message protocol between the paper and the worker (`add` / `remove` / `change` / `reset`) and an "awaiting-update" visual state on links while routing is in flight.
+
+```bash
+npm run start-web-worker   # or: npm run build-web-worker
+```
+
+### Web Worker (performance)
+
+Stress-tests the worker-based setup with larger diagrams. A dropdown switches between two pre-built flowchart-style graphs of very different sizes (the bigger one has several hundred elements); each switch resets the worker so every run starts from clean libavoid state. Uses dedicated shapes from `src/web-worker-perf/shapes.js` and has no editing affordances — click on blank or a cell to fit the paper, drag on blank to rubber-band zoom to a region. Routing time is logged to the console via `console.time('worker routed')`.
+
+```bash
+npm run start-web-worker-perf   # or: npm run build-web-worker-perf
 ```
